📝 (content): update video metadata and add captions for better accessibility and SEO
The video metadata is updated to reflect a more engaging and descriptive title and slug, enhancing discoverability and user engagement. Captions are added to improve accessibility for viewers with hearing impairments and to provide a better understanding of the video's content. This change also aids in SEO, making the video more likely to appear in search results. The addition of aliases ensures that previous links remain functional, preventing broken links and maintaining user access.

📝 (resources): add English captions and metadata for YouTube videos

Introduce English subtitles and metadata for YouTube videos to enhance accessibility and improve user engagement. This change ensures that non-native speakers and those with hearing impairments can better understand the content. Additionally, it provides SEO benefits by making the video content more searchable and indexable.

✨ (content): add new video resources on transitioning from monolith to microservices

Introduce new video content focused on transitioning from monolithic
architectures to microservices. This addition aims to provide valuable
insights and strategies for CTOs and engineering leads facing the
challenges of modernizing legacy systems. The content emphasizes
practical guidance, strategic decomposition, and the importance of
team dynamics in achieving a successful transition. By offering a
step-by-step approach, the resources aim to enhance system flexibility
and scalability, aligning with business objectives and improving
delivery capabilities.

📝 (youtube.json): update video metadata and descriptions for playlist items

Update the metadata and descriptions of YouTube playlist items to reflect the latest content and ensure accuracy. This includes changes to video titles, descriptions, publication dates, and thumbnail URLs. The updates are necessary to keep the playlist current and relevant, providing viewers with the most up-to-date information and enhancing the overall user experience.